pyproject.toml
[project]
name = "wordle-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "psycopg2>=2.9.10",
    "rich>=14.0.0",
    "sqlalchemy>=2.0.41",
]

[project.scripts]
bot = "wordle_discord_bot:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["mypy>=1.16.1", "python-dotenv>=1.1.1", "ruff>=0.12.1"]

[tool.ruff]
lint.extend-select = ["I"] # enable import sorting
